Add integration testing functionality and documentation
This adds a new package, Google.Cloud.Functions.Invoker.Testing,
containing a test server - as well as methods in EntryPoint to just
create an IHostBuilder where that's more desirable.

The documentation here is roughly the level and style of
documentation I'm hoping to use for other features, such as
dependency injection.