Feat: Rearchitecture of Functions Framework hosting.
This is now much more idiomatic of ASP.NET Core, and also provides
significantly richer customization via FunctionsStartup classes.

Tests cover the same scenarios as before (and a few more).

This commit does not include documentation, or samples of additional
functionality (e.g. FunctionTestServerBuilder). However, it is
complete in itself.