Include Hosting package in Middleware example
The build.sh script didn't detect the problem because we always
build with LocalFunctionsFramework set to true - which makes sense
in terms of allowing us to write examples for unreleased framework
features, but hides this sort of thing. I think it's a pretty rare
problem - I don't think we need to change the build.