Make FunctionsStartupAttribute apply to classes as well as assemblies
We guess which function will be the target, and use that to find startups. It's just possible (but very unlikely) that this will prove incorrect, and that we'll have the "wrong" set of startups afterwards... the next commit will detect and throw if that happens.

This means we can use the same attribute for both tests and production code, which is nice - decorate the test class with FunctionsStartupAttribute to specify the startup classes for that test.