test: Creates a buildpack validation action
This will ensure that any changes to the Functions Framework are
compatible with the buildpack. In order to use the latest source
code, it builds the Functions Framework locally and includes the
package file in a temporary copy of the conformance tests, along
with a nuget.config file to use that local package.